requirements.txt
google-generativeai
gspread
oauth2client
python-telegram-bot
